backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# LLM - Gemini (does OCR + extraction!)
google-generativeai==0.8.3

# Financial Calculations
pandas==2.2.3
numpy==2.1.2

# Image Processing (for basic file handling)
Pillow==10.4.0
pdf2image==1.17.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
